+++ aimake/plugins/__init__.py
@@ -1,5 +1,7 @@
 """Plugins package."""
 
 from aimake.plugins.base import AimakePlugin, PluginManager
+from aimake.plugins.huggingface import HuggingFacePlugin
+from aimake.plugins.loader import load_plugins
 
-__all__ = ["AimakePlugin", "PluginManager"]
+__all__ = ["AimakePlugin", "HuggingFacePlugin", "PluginManager", "load_plugins"]